+++ Tests/SentryTests/Helper/SentryFileManagerTests.swift
@@ -2,8 +2,44 @@
 @testable import Sentry.SentryOptions
 import XCTest
 
+// Even if we don't run this test below OSX 10.12 we expect the actual
+// implementation to be thread safe.
+@available(OSX 10.12, *)
 class SentryFileManagerTests: XCTestCase {
     
+    private class Fixture {
+        
+        let session = SentrySession(releaseName: "1.0.0")
+        let sessionEnvelope: SentryEnvelope
+        
+        let sessionUpdate: SentrySession
+        let sessionUpdateEnvelope: SentryEnvelope
+        
+        let expectedSessionUpdate: SentrySession
+        
+        let queue = DispatchQueue(label: "SentryFileManagerTests", qos: .utility, attributes: [.concurrent, .initiallyInactive])
+        let group = DispatchGroup()
+        
+        init() {
+            sessionEnvelope = SentryEnvelope(session: session)
+            
+            sessionUpdate = session.copy() as! SentrySession
+            sessionUpdate.incrementErrors()
+            
+            let event = Event()
+            let items = [SentryEnvelopeItem(session: sessionUpdate), SentryEnvelopeItem(event: event)]
+            sessionUpdateEnvelope = SentryEnvelope(id: event.eventId, items: items)
+            
+            let sessionUpdateCopy = sessionUpdate.copy() as! SentrySession
+            // We need to serialize in order to set the timestamp and the duration
+            expectedSessionUpdate = SentrySession(jsonObject: sessionUpdateCopy.serialize())
+            // We can only set the init flag after serialize, because the duration is not set if the init flag is set
+            expectedSessionUpdate.setFlagInit()
+        }
+        
+    }
+    
+    private var fixture: Fixture!
     private var sut: SentryFileManager!
     private var currentDateProvider: TestCurrentDateProvider!
 
@@ -12,6 +48,8 @@ class SentryFileManagerTests: XCTestCase {
         do {
             currentDateProvider = TestCurrentDateProvider()
             CurrentDate.setCurrentDateProvider(currentDateProvider)
+            
+            fixture = Fixture()
 
             sut = try SentryFileManager(dsn: TestConstants.dsn, andCurrentDateProvider: currentDateProvider)
 
@@ -90,6 +128,18 @@ class SentryFileManagerTests: XCTestCase {
         for _ in 0...100 {
             sut.store(TestConstants.envelope)
         }
+        
+        let events = sut.getAllEnvelopes()
+        XCTAssertEqual(events.count, 100)
+    }
+    
+    func testDefaultMaxEnvelopesConcurrent() {
+        for _ in 0...1_000 {
+            storeAsync(envelope: TestConstants.envelope)
+        }
+        fixture.queue.activate()
+        fixture.group.wait()
+        
         let events = sut.getAllEnvelopes()
         XCTAssertEqual(events.count, 100)
     }
@@ -101,6 +151,69 @@ class SentryFileManagerTests: XCTestCase {
         }
         let events = sut.getAllEnvelopes()
         XCTAssertEqual(events.count, 15)
+    }
+    
+    func testMigrateSessionInit_SessionUpdateIsLast() {
+        sut.store(fixture.sessionEnvelope)
+        // just some other session
+        sut.store(SentryEnvelope(session: SentrySession(releaseName: "1.0.0")))
+        for _ in 0...97 {
+            sut.store(TestConstants.envelope)
+        }
+        sut.store(fixture.sessionUpdateEnvelope)
+        
+        assertSessionInitMoved(sut.getAllEnvelopes().last!)
+        assertSessionEnvelopesStored(count: 2)
+    }
+
+    func testMigrateSessionInit_SessionUpdateIsSecond() {
+        sut.store(fixture.sessionEnvelope)
+        sut.store(fixture.sessionUpdateEnvelope)
+        for _ in 0...98 {
+            sut.store(TestConstants.envelope)
+        }
+
+        assertSessionInitMoved(sut.getAllEnvelopes().first!)
+        assertSessionEnvelopesStored(count: 1)
+    }
+    
+    func testMigrateSessionInit_IsInMiddle() {
+        sut.store(fixture.sessionEnvelope)
+        for _ in 0...50 {
+            sut.store(TestConstants.envelope)
+        }
+        sut.store(fixture.sessionUpdateEnvelope)
+        for _ in 0...48 {
+            sut.store(TestConstants.envelope)
+        }
+
+        assertSessionInitMoved(sut.getAllEnvelopes()[50])
+        assertSessionEnvelopesStored(count: 1)
+    }
+    
+    func testMigrateSessionInit_NoOtherSessionUpdate() {
+        sut.store(fixture.sessionEnvelope)
+        sut.store(fixture.sessionUpdateEnvelope)
+        for _ in 0...99 {
+            sut.store(TestConstants.envelope)
+        }
+        
+        assertSessionEnvelopesStored(count: 0)
+    }
+    
+    /**
+     * We need to deserialize every envelope and check if it contains a session.
+     */
+    func testMigrateSessionInit_WorstCasePerformance() {
+        sut.store(fixture.sessionEnvelope)
+        sut.store(fixture.sessionUpdateEnvelope)
+        for _ in 0...97 {
+            sut.store(TestConstants.envelope)
+        }
+        
+        measure {
+            sut.store(TestConstants.envelope)
+        }
     }
 
     func testGetAllEnvelopesAreSortedByDateAscending() {
@@ -179,6 +292,14 @@ class SentryFileManagerTests: XCTestCase {
 
         XCTAssertEqual(0, sut.getAllEnvelopes().count)
     }
+    
+    private func storeAsync(envelope: SentryEnvelope) {
+        fixture.group.enter()
+        fixture.queue.async {
+            self.sut.store(envelope)
+            self.fixture.group.leave()
+        }
+    }
 
     private func storeEvent() {
         do {
@@ -193,6 +314,25 @@ class SentryFileManagerTests: XCTestCase {
     private func assertEventFolderDoesntExist() {
         XCTAssertFalse(FileManager.default.fileExists(atPath: sut.eventsPath),
                         "Folder for events should be deleted on init: \(sut.eventsPath)")
+    }
+    
+    private func assertSessionInitMoved(_ actualSessionFileContents: SentryFileContents) {
+        let actualSessionEnvelope = SentrySerialization.envelope(with: actualSessionFileContents.contents)
+        XCTAssertEqual(2, actualSessionEnvelope?.items.count)
+
+        let actualSession = SentrySerialization.session(with: actualSessionEnvelope?.items[1].data ?? Data())
+        XCTAssertNotNil(actualSession)
+
+        XCTAssertEqual(fixture.expectedSessionUpdate, actualSession)
+    }
+    
+    private func assertSessionEnvelopesStored(count: Int) {
+        let fileContentsWithSession = sut.getAllEnvelopes().filter { envelopeFileContents in
+            let envelope = SentrySerialization.envelope(with: envelopeFileContents.contents)
+            return envelope?.items.filter { item in item.header.type == SentryEnvelopeItemTypeSession }.count != 0
+        }
+        
+        XCTAssertEqual(count, fileContentsWithSession.count)
     }
 
     private func advanceTime(bySeconds: TimeInterval) {