+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## unreleased
 
+- feat: Migrate session init for stored envelopes #693
+
 - fix: Rate limiting for cached envelope items #685
 - feat: Errors and sessions in the same envelope #686
 - feat: Implement NSCopying for SentrySession #683